+++ site/content/changelog.md
@@ -8,12 +8,11 @@ weight: 2
 Changes to this project will be documented here.
 
 ### [0.3.5] - Unreleased
-
-  * Reminder to re-minify js in app.html
   
   * Lots of code refactoring
   * Maybe slight performance improvement
-  * Probably break IE11 with ES6 features
+  * Add babel to not break IE11 with ES6 features
+  * Better minifier
   * Drawing name validates on input instead of output
   * Removed tooltip and snap options from export/import file
   * Save options across sessions with local storage